Eliminar archivos innecesarios y agregar configuración de CI/CD para despliegue en AWS
