+++ .github/main.yml
@@ -19,7 +19,7 @@ jobs:
         key: ${{ secrets.PRIVATE_KEY }}
         port: ${{ secrets.PORT }}
         script: |
-          mkdir -p /home/ubuntu/web-usuario
+          mkdir -p /home/ubuntu/web-nueva
 
     - name: Sincronizar archivos con el servidor
       uses: burnett01/rsync-deployments@7.0.2
@@ -28,7 +28,7 @@ jobs:
         path: "./" # Directorio de origen (los archivos del repositorio)
         remote_host: ${{ secrets.IP_SERVER }} # Dirección IP del servidor
         remote_user: ${{ secrets.USERNAME }} # Usuario SSH en el servidor
-        remote_path: "/home/ubuntu/web-usuario/" # Ruta de destino en el servidor
+        remote_path: "/home/ubuntu/web-nueva/" # Ruta de destino en el servidor
         remote_key: ${{ secrets.PRIVATE_KEY }} # Clave SSH para autenticar
         remote_port: ${{ secrets.PORT }} # Puerto SSH ( si es distinto del predeterminado 22 )
 
@@ -40,6 +40,6 @@ jobs:
         key: ${{ secrets.PRIVATE_KEY }}
         port: ${{ secrets.PORT }}
         script: |
-          cd /home/ubuntu/web-usuario
+          cd /home/ubuntu/web-nueva
           sudo docker compose down
           sudo docker compose up -d --build
