--- docker-compose.yml
@@ -1,8 +1,0 @@
-services:
-  java:
-    build:
-      context: .
-      dockerfile: dockerfile
-    ports:
-      - "8081:8081"
-    command: [ "java", "-jar", "app.jar" ]
