--- src/main/java/com/usuario/usuario/config/AppConfig.java
@@ -1,14 +1,0 @@
-package com.usuario.usuario.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class AppConfig {
-    @Bean
-    public RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
-
-}
