--- src/main/java/com/usuario/usuario/config/SwaggerConfig.java
@@ -1,19 +1,0 @@
-package com.usuario.usuario.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-
-@Configuration
-public class SwaggerConfig {
-    @Bean
-    public OpenAPI openAPI() {
-        return new OpenAPI().info(new Info()
-                .title("Proyecto de usuarios")
-                .version("1.0")
-                .description("Este es mi microservicio"));
-    }
-
-}
