+++ src/main/java/com/usuario/usuario/model/Usuario.java
@@ -17,6 +17,7 @@ import lombok.NoArgsConstructor;
 
 public class Usuario {
 
+    private Long idProducto;
     private Long idUsuario;
 
     @NotNull
