--- src/main/resources/application-dev.properties
@@ -1,24 +1,0 @@
-# Configuración del servidor para desarrollo
-server.port=8081
-
-# Configuración de la base de datos H2 para desarrollo local
-spring.datasource.url=jdbc:h2:file:./data/usuario_dev
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=
-
-# Configuración de JPA / Hibernate para H2
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.H2Dialect
-
-# Configuración de H2 Console
-spring.h2.console.enabled=true
-spring.h2.console.path=/h2-console
-
-# Nombre del proyecto
-spring.application.name=usuario-service-dev
-
-# Logging para desarrollo
-logging.level.org.springframework.web=INFO
-logging.level.com.usuario.usuario=DEBUG
