+++ src/main/resources/application.properties
@@ -14,4 +14,3 @@ spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
 
 # Nombre del proyecto
 spring.application.name=usuario-service
-
