+++ src/test/java/com/usuario/usuario/TestUsuario.java
@@ -1,21 +1,17 @@
 package com.usuario.usuario;
 
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.*;
+
 import java.util.Optional;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import static org.mockito.ArgumentMatchers.any;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import static org.mockito.Mockito.doNothing;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 import org.mockito.MockitoAnnotations;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
 
 import com.usuario.usuario.model.Usuario;
 import com.usuario.usuario.model.entity.UsuarioEntity;
@@ -23,7 +19,6 @@ import com.usuario.usuario.repository.UsuarioRepository;
 import com.usuario.usuario.service.UsuarioService;
 
 @SpringBootTest
-@ActiveProfiles("test")
 public class TestUsuario {
 
     @Mock
