+++ src/test/java/com/usuario/usuario/UsuarioApplicationTests.java
@@ -2,10 +2,8 @@ package com.usuario.usuario;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest
-@ActiveProfiles("test")
 class UsuarioApplicationTests {
 
 	@Test
