--- src/test/resources/application-test.properties
@@ -1,24 +1,0 @@
-# Configuración del servidor para pruebas
-server.port=8081
-
-# Configuración de la base de datos H2 en memoria para pruebas
-spring.datasource.url=jdbc:h2:mem:testdb
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=password
-
-# Configuración de JPA / Hibernate para H2
-spring.jpa.hibernate.ddl-auto=create-drop
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.H2Dialect
-
-# Configuración de H2 Console (para debugging si es necesario)
-spring.h2.console.enabled=true
-spring.h2.console.path=/h2-console
-
-# Nombre del proyecto
-spring.application.name=usuario-service-test
-
-# Logging para pruebas
-logging.level.org.springframework.web=DEBUG
-logging.level.com.usuario.usuario=DEBUG
